Add description and how to use
The script converts the EXIF info into a usable format that can be cut and pasted in to Google/Bing Maps. The format is as follows: 40 45' 36.00" N, 73 59' 2.40 W  

From the Terminal execute the script and image you would like to be read; the image file has to being the same folder as the script unless you input the path. (If you plan on doing a lot of images I suggest creating a folder to co-locate the script and files).